TS: ProteinListFilters/CurationFilter
